change price to float

+++ Vinci Stock eXchange/execution/src/main/java/be/vinci/ipl/execution/ExecutionService.java
@@ -48,7 +48,7 @@ public class ExecutionService {
    * @param ticker the ticker's identifier.
    * @param newPrice the ticker's new price.
    */
-  public void updatePrice(String ticker, int newPrice){
+  public void updatePrice(String ticker, float newPrice){
     priceProxy.updateOne(ticker, String.valueOf(newPrice));
   }
